moving the chat loader to router.ts and fixing the case when the user does not have friends
