+++ client/src/pages/Chat.module.scss
@@ -125,3 +125,10 @@
         }
     }
 }
+
+.noFriends {
+    color: #888;
+    font-style: italic;
+    padding: 1rem;
+    text-align: center;
+}
