+++ client/src/router.ts
@@ -14,7 +14,7 @@ import { Notifications } from "./pages/Notifications";
 import { fetchNotifications } from "./models/notifications";
 import { PauseOctagon } from "lucide-react";
 import { getComments } from "./models/comments";
-import { Chat, chatLoader } from "./pages/Chat";
+import { Chat } from "./pages/Chat";
 
 export const router = createBrowserRouter([
     {
@@ -96,7 +96,25 @@ export const router = createBrowserRouter([
             },
             { path: "/chat", 
                 Component: Chat,
-                loader: chatLoader
+                loader: async () => {
+                    const userId = getLoggedInUserId();;
+                    const username = getLoggedInUserName();
+
+                    if (!userId || !username || userId === "Guest") {
+                        console.warn("User is not logged in. Redirecting to login");
+                        return redirect("/login");
+                    }
+
+                    try {
+                        const res = await fetch(`http://localhost:5050/chat/friends/${userId}`);
+                        const friends = await res.json();
+                
+                        return { userId, username, friends };
+                    } catch (error) {
+                        console.error("Failed to load friends:", error);
+                        throw new Response("Failed to load friends", { status: 500 });
+                    }
+                }
             }    
         ]
     }
